requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
backports.functools-lru-cache==1.5
boto3==1.9.83
botocore==1.12.83
certifi==2018.11.29
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.1.0
docutils==0.14
glob2==0.6
idna==2.7
jaraco.functools==1.20
Jinja2==2.10.1
jmespath==0.9.3
MarkupSafe==1.1.0
more-itertools==4.3.0
multidict==4.5.2
ordereddict==1.1
portend==2.3
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1.1
requests==2.20.1
s3transfer==0.1.13
selenium==3.141.0
six==1.11.0
tempora==1.14
urllib3==1.24.2
yarl==1.2.6
zc.lockfile==1.4
